requirements.txt
Django==4.2.4
python-slugify==8.0.1
